restructured the unit tests
